Mv raw methods to DB class
